+++ src/common/components/layout/common-layout.tsx
@@ -1,9 +1,9 @@
 import { Outlet } from 'react-router';
 
 import { PageUrlType } from '@/common/constants/path';
-import { generateHeaderContents } from '@/common/utils/generateHederContents';
+import { useHederContents } from '@/common/hooks/use-heder-contents';
 
-import HeaderNavigationLayout from '../header-navigation/header-navigation-layout';
+import HeaderNavigationLayout from './header-navigation/header-navigation-layout';
 
 import * as styles from './common-layout.styles';
 
@@ -13,7 +13,7 @@ interface CommonLayoutProps {
 }
 
 function CommonLayout({ isHeader, pageUrl }: CommonLayoutProps) {
-  const { left, middle, right } = generateHeaderContents(pageUrl);
+  const { left, middle, right } = useHederContents(pageUrl);
 
   return (
     <div css={styles.container(pageUrl)}>
